args.py: Select multiple scan channels
Remove the integer restriction for selecting a channel.
Doing so we pass the argument from wifite directly to
airodump which does support the format 1,3-7,11-13.

Added a regex which checks the channel argument value
and raises a ValueError exception in case an invalid
values is sent. Sending this value as in to airodump
would result in airodump crashing with no explanation.

Signed-off-by: Radu Nicolau <radunicolau102@gmail.com>