Restrict exceptions caught in functions
The issue was that sending Ctrl+C would not be caught by the
try-except clause inside the scanner if the keyboard interrupt
was sent while inside inner try-except blocks. This would
result in airodump crashing and the Scanner.find_targets()
function returning, instead of prompting the user with the
choice to stop the attack.
This could be the possible cause for issue #123

TODO: catch specific exceptions where possible

Signed-off-by: Radu Nicolau <radunicolau102@gmail.com>