Print more helpful airodump failure reason
* Removed exception in case target list is empty.

Normal scan mode will not be affected as it will
try to attack an empty list. The reason I removed
the exception is because in infinite attack mode it
is possible to press Ctrl+C too fast, resulting in 0
targets, yet it is not an error.

* Added error messsage in case airodump crashes.

The output will be the exit status code and the command
which was ran. Because airodump also prints to stderr
the useful output, we cannot realiably print error
messages from airodump.

* Added chipset/driver information link

Despite the fact that the link says the information hasn't
been updated, edits are still being made and the existing
chipsets will not change their specs. Coupled with driver
information, this link seems very useful for self debugging.

Signed-off-by: Radu Nicolau <radunicolau102@gmail.com>