Implement --power option
Only display and attack APs that had at least <power> during the
current scan.

Because the power is not very stable it was possible for a target
to drop below the power level and be deleted from the target list.
This would in turn cause the updated targets not to overwrite the
existing list, which would cause wifite to output an unnecessary
amount of lines. The workaround is to compare the maximum amount
of power an AP had.

Signed-off-by: Radu Nicolau <radunicolau102@gmail.com>